% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{apply_one_hot_encode_new_data}
\alias{apply_one_hot_encode_new_data}
\title{Apply One-Hot Encoding to New Data}
\usage{
apply_one_hot_encode_new_data(new_data, levels_list)
}
\arguments{
\item{new_data}{The dataframe containing new observations.}

\item{levels_list}{A list detailing the categories for each categorical column from the original dataset.}
}
\value{
The transformed dataset with original dataset's structure.
}
\description{
Encodes new data to fit the one-hot encoding scheme used on the training data. It ensures that only previously seen categories are present in the new data.
}
\examples{
train_data <- data.frame(Color = factor(c("Red", "Green", "Blue")),
                         Shape = factor(c("Square", "Circle", "Triangle")),
                         Value = c(10, 20, 30))
encoded_train_data <- perform_one_hot_encoding(train_data, return_levels = TRUE)
levels_list <- encoded_train_data$levels
new_data <- data.frame(Color = factor(c("Red", "Green")), Shape = factor(c("Square", "Circle")))
encoded_new_data <- apply_one_hot_encode_new_data(new_data, levels_list)
}
