% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{calculate_mode}
\alias{calculate_mode}
\title{Calculate Mode}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{numerical list or vector.}
}
\value{
The mode of 'x'.
}
\description{
Finds the most frequent value in a vector, excluding NA.
}
