% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{convert_and_clean_columns}
\alias{convert_and_clean_columns}
\title{Convert and Clean Dataframe Columns}
\usage{
convert_and_clean_columns(dataframe)
}
\arguments{
\item{dataframe}{Dataframe to be processed.}
}
\value{
Dataframe with cleaned and potentially converted columns.
}
\description{
Converts character and logical columns to factors, binary numeric columns to factors,
and splits Date columns into year, month, and day. Drops constant columns.
}
\examples{
df <- data.frame(a = c(1, 1, 1), b = c(TRUE, FALSE, TRUE),
  date = as.Date(c('2020-01-01', '2020-01-02', '2020-01-03')))
cleaned_df <- convert_and_clean_columns(df)
}
