% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_learning.R
\name{do_stacking}
\alias{do_stacking}
\title{Stacking Ensemble}
\usage{
do_stacking(
  predictions_list,
  y,
  is_binary_classification,
  classification_threshold
)
}
\arguments{
\item{predictions_list}{A list of prediction vectors from base models.}

\item{y}{The true response values.}

\item{is_binary_classification}{Boolean indicating if the task is binary classification.}

\item{classification_threshold}{Threshold for binary classification.}
}
\value{
Predictions from the stacking model.
}
\description{
Combines predictions using stacking, where a meta-model is trained on the predictions of base models.
}
\examples{
predictions1 <- runif(10, 0, 1)
predictions2 <- runif(10, 0, 1)
y <- rbinom(10, 1, 0.5)  # For binary classification
# y <- runif(10, 0, 1)  # For regression
stacked_predictions <- do_stacking(list(predictions1, predictions2), y, TRUE, 0.5)
}
