% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{feature_selection_method_assumptions}
\alias{feature_selection_method_assumptions}
\title{This function checks if the response variable \code{y} is appropriate for the chosen feature selection method.}
\usage{
feature_selection_method_assumptions(y, method)
}
\arguments{
\item{y}{The response variable}

\item{method}{The feature selection method being used ("logistic" or "correlation" or "random_forest" or "auto").}
}
\value{
No return value; stops with an error message if assumptions are not met.
}
\description{
This function checks if the response variable \code{y} is appropriate for the chosen feature selection method.
}
\examples{
# For a binary response variable
feature_selection_method_assumptions(c(0, 1, 0, 1), "logistic")

# For a continuous response variable
feature_selection_method_assumptions(runif(10), "correlation")

}
