% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{find_optimal_k}
\alias{find_optimal_k}
\title{Find the optimal value for the number of top predictors (k).}
\usage{
find_optimal_k(n_predictors, n_records)
}
\arguments{
\item{n_predictors}{The total number of predictors in the dataset.}

\item{n_records}{The total number of records in the dataset.}
}
\value{
The optimal value for the number of top predictors (k).
}
\description{
This function calculates an optimal value for the number of top predictors (k) based on the size of the dataset.
}
\examples{
find_optimal_k(100, 200)
# Returns 50

}
