% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_model}
\alias{fit_model}
\title{Fit Various Types of Models}
\usage{
fit_model(X, y, model_type, params = list())
}
\arguments{
\item{X}{Input matrix of predictors.}

\item{y}{Response variable.}

\item{model_type}{Type of model to fit.}

\item{params}{List of additional parameters for model fitting.}
}
\value{
Fitted model object.
}
\description{
Fits a model based on input data, model type, and parameters.
Supports logistic regression, linear regression, lasso, ridge, elastic net, and random forest.
}
\examples{
n <- 20
X <- matrix(rnorm(n * 10), ncol = 10)
y <- rnorm(100, 1, 0.5)
fit_model(X, y, "linear_regression")
n = 10
X1 <- rnorm(n, mean = 10, sd = 2)
X2 <- runif(n, min = 0, max = 1)
X3 <- rbinom(n, size = 1, prob = 0.3)
X <- data.frame(X1, X2, X3)
true_coeffs <- c(2, -3, 1)
y <- X1 * true_coeffs[1] + X2 * true_coeffs[2] + X3 * true_coeffs[3] + rnorm(n, mean = 0, sd = 1)
params = list(lamda = 0.05)
fit_model(X, y, model_type="elastic_net", params=params)

}
