% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tuning_cv.R
\name{handle_glmnet_tuning}
\alias{handle_glmnet_tuning}
\title{Tune glmnet Models - lasso, ridge or elastic_net}
\usage{
handle_glmnet_tuning(X_train, y_train, alpha, user_params, cv_folds)
}
\arguments{
\item{X_train}{Training features.}

\item{y_train}{Training labels.}

\item{alpha}{Elastic net mixing parameter.}

\item{user_params}{Parameters list, including lambda.}

\item{cv_folds}{Number of cross-validation folds.}
}
\value{
List with best lambda and alpha.
}
\description{
Tunes the lambda parameter for glmnet models using cross-validation.
}
\examples{
# Example usage:
n = 100
X_train <- data.frame(
 feature_num1 = rnorm(n, mean = 5, sd = 2),  # Numerical feature
 feature_num2 = rnorm(n, mean = 55, sd = 42),  # Numerical feature
 feature_num3 = rnorm(n, mean = 544, sd = 52),  # Numerical feature
 feature_num4 = rnorm(n, mean = 59, sd = 16)  # Numerical feature
 )
y_train = sample(0:1, n, replace=TRUE)
best_params <- handle_glmnet_tuning(X_train, y_train, 1, list(lambda = c(0.01, 0.1)), 10)

}
