% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values}
\usage{
impute_missing_values(column, method = "mean")
}
\arguments{
\item{column}{to impute.}

\item{method}{imputation method ('mean', 'median', 'mode').}
}
\value{
imputed column
}
\description{
Replaces missing values in a column using the specified method: mean, median, or mode.
}
\examples{
impute_missing_values(c(1, 2, NA, 4), method = "mean")

}
