% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{is_k_top_predictors_valid}
\alias{is_k_top_predictors_valid}
\title{Check if the number of top predictors (k) is valid.}
\usage{
is_k_top_predictors_valid(k, x)
}
\arguments{
\item{k}{Number of top predictors you want to use. Must be a positive whole number.}

\item{x}{Dataset to check the number of predictors against. Can't have more predictors than variables.}
}
\value{
TRUE if 'k' is valid, FALSE otherwise.
}
\description{
This function checks if your chosen number of top predictors (k) is sensible for the dataset.
}
\examples{
is_k_top_predictors_valid(5, iris) # Should be TRUE if iris has 5 or more variables.
is_k_top_predictors_valid(0, iris) # Will be FALSE, 0 is not a valid number.

}
