% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{log_message}
\alias{log_message}
\title{Log a Message}
\usage{
log_message(msg)
}
\arguments{
\item{msg}{The message string to log.}
}
\description{
Conditionally logs a message to the R console, depending on the global \code{enable_logs} flag.
}
\details{
If \code{enable_logs} is TRUE, the message will be prefixed with "Message: " and printed.
}
