% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_learning.R
\name{majority_voting}
\alias{majority_voting}
\title{Majority Voting for classification predictions}
\usage{
majority_voting(predictions_list, classification_threshold)
}
\arguments{
\item{predictions_list}{A list of binary prediction vectors from different models.}

\item{classification_threshold}{Threshold for converting probabilities to class labels.}
}
\value{
A vector of predictions based on majority voting.
}
\description{
Combines binary class predictions from multiple models using majority voting.
}
\examples{
predictions1 <- rbinom(10, 1, 0.5)
predictions2 <- rbinom(10, 1, 0.5)
voted_predictions <- majority_voting(list(predictions1, predictions2), 0.5)
}
