% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{perform_bagging}
\alias{perform_bagging}
\title{Performs Bagging operation}
\usage{
perform_bagging(X, y, model_type, params, B = 100, classification_threshold)
}
\arguments{
\item{X}{The predictor variables as a matrix or dataframe.}

\item{y}{The response variable vector.}

\item{model_type}{The type of model to use ('random_forest', 'linear_regression', 'logistic_regression', 'lasso', 'ridge', 'elastic_net').}

\item{params}{Parameters specific to the chosen model type.}

\item{B}{The number of bootstrap samples to generate.}

\item{classification_threshold}{The threshold for classifying observations as 1 in binary classification tasks.}
}
\value{
A list containing three elements:
- \code{predictions}: the predicted values or classes,
- \code{variable_importance}: the importance scores of the predictor variables,
- \code{metrics}: performance metrics of the model (accuracy, sensitivity, specificity, precision for classification; MSE, RMSE, MAE, R-squared for regression).
}
\description{
Applies a bagging ensemble method to either regression or binary classification tasks
using specified model types and model parameters.
}
\examples{
n = 100
data <- data.frame(
    feature_num1 = rnorm(n, mean = 5, sd = 2),  # Numerical feature
    feature_num2 = rnorm(n, mean = 55, sd = 42),  # Numerical feature
    feature_num3 = rnorm(n, mean = 544, sd = 52),  # Numerical feature
    feature_num4 = rnorm(n, mean = 59, sd = 16),  # Numerical feature
    # target = sample(0:1, n, replace = TRUE) # For classification
    target = rnorm(sample(0:1, n, replace = TRUE)) # For regression
  )
X <- data[, -which(names(data) == "target")]
y <- data[["target"]]
model_type <- "lasso"
params <- list(ntree = 100)
classification_threshold <- 0.5  # Relevant for binary classification
results <- perform_bagging(X, y, model_type, params, B = 100, classification_threshold)
print(results)

}
