% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{perform_one_hot_encoding}
\alias{perform_one_hot_encoding}
\title{Perform One-Hot Encoding on Factor Columns}
\usage{
perform_one_hot_encoding(dataframe, return_levels = FALSE)
}
\arguments{
\item{dataframe}{A dataframe containing the data to be one-hot encoded.}

\item{return_levels}{A logical value indicating whether to return the levels of the
factor columns along with the encoded dataframe. Default is FALSE.}
}
\value{
If \code{return_levels} is FALSE, it will only return the changed dataframe.
If TRUE, it will return a list with the dataframe and the levels of each factor column.
}
\description{
This function performs one-hot encoding on factor columns in a dataframe. It replaces
each factor column with a set of binary columns, one for each level of the factor.
The function can optionally return the levels of the factor columns along with the
encoded dataframe.
}
\examples{
# Load an example dataset
data(iris)

# One-hot encode the 'Species' column
encoded_data <- perform_one_hot_encoding(iris)

# One-hot encode and return the levels
encoded_data_with_levels <- perform_one_hot_encoding(iris, return_levels = TRUE)
encoded_iris <- encoded_data_with_levels$data
species_levels <- encoded_data_with_levels$levels

}
