% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Data Preprocessing Function}
\usage{
preprocess_data(
  y,
  x,
  models,
  feature_selection_method = NULL,
  k = NULL,
  drop_missing_records = TRUE,
  fill_missing_method = "mean",
  scale_data = FALSE,
  remove_outliers = FALSE,
  seed = 123
)
}
\arguments{
\item{y}{Response variable.}

\item{x}{Predictor dataframe}

\item{models}{List of model names ('logistic_regression','linear_regression','lasso','ridge','elastic_net','random_forest)}

\item{feature_selection_method}{Method for feature selection.}

\item{k}{Number of top predictors.}

\item{drop_missing_records}{Boolean; flag that indicates whether to drop the records (TRUE by default)}

\item{fill_missing_method}{Method for imputing missing values ('mean' by default).}

\item{scale_data}{Boolean; flag that indicates whether to scale continous data.}

\item{remove_outliers}{Boolean; flag that indicates whether to remove outliers (FALSE by default).}

\item{seed}{Seed for reproducibility (123 by default).}
}
\value{
A list containing preprocessed dataset x and additional information like scaling parameters, one-hot encoding levels, and names of selected top K features.
}
\description{
Performs comprehensive preprocessing on any given dataset, including handling missing values,
converting and cleaning columns, removing outliers, encoding categorical variables,
selecting top K predictors, and scaling the data (optional).
}
\examples{
n = 100
data <- data.frame(
    feature_num1 = rnorm(n, mean = 5, sd = 2),  # Numerical feature
    feature_num2 = rnorm(n, mean = 55, sd = 42),  # Numerical feature
    feature_num3 = rnorm(n, mean = 544, sd = 52),  # Numerical feature
    feature_num4 = rnorm(n, mean = 59, sd = 16),  # Numerical feature
    feature_cat = sample(c("Category1", "Category2", "Category3"), 
n, replace = TRUE),  # Categorical feature
    target = sample(0:1, n, replace = TRUE)  # Binary target variable
  )
x <- data[, -which(names(data) == "target")]
y <- data[["target"]]
models <- c("logistic_regression","lasso")
preprocessed <- preprocess_data(y, x, models, scale_data = TRUE, remove_outliers = TRUE)

}
