% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{remove_outliers_iqr}
\alias{remove_outliers_iqr}
\title{Remove Outliers Using IQR Method}
\usage{
remove_outliers_iqr(dataframe, y)
}
\arguments{
\item{dataframe}{Dataframe containing predictor variables.}

\item{y}{Response variable.}
}
\value{
A list with updated 'y' (response variable) and 'dataframe' excluding rows identified as outliers.
}
\description{
Identifies and removes outliers from a dataframe based on the Interquartile Range (IQR) method.
}
\examples{
df <- data.frame(a = rnorm(100), b = rnorm(100))
y <- rnorm(100)
cleaned_data <- remove_outliers_iqr(df, y)

}
