% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{select_top_k_predictors}
\alias{select_top_k_predictors}
\title{Select Top K Predictors}
\usage{
select_top_k_predictors(x, y, k, method = "auto")
}
\arguments{
\item{x}{Predictor dataset.}

\item{y}{Response variable.}

\item{k}{Number of top predictors to select.}

\item{method}{Method for selection ('auto', 'logistic', 'correlation', 'random_forest'). 'auto' detects based on 'y'.}
}
\value{
Names of the top K predictors.
}
\description{
Selects top K predictors from 'x' for the response variable 'y' using specified or auto-detected method.
}
\examples{
n = 100
data <- data.frame(
    feature_num1 = rnorm(n, mean = 5, sd = 2),  # Numerical feature
    feature_num2 = rnorm(n, mean = 55, sd = 42),  # Numerical feature
    feature_num3 = rnorm(n, mean = 544, sd = 52),  # Numerical feature
    feature_num4 = rnorm(n, mean = 59, sd = 16),  # Numerical feature
    target = sample(0:1, n, replace = TRUE)  # Binary target variable
  )
x <- data[, -which(names(data) == "target")]
y <- data[["target"]]
select_top_k_predictors(x = x, y = y, k = 3, method = "auto")

}
