% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_learning.R
\name{soft_averaging}
\alias{soft_averaging}
\title{Soft Averaging Ensemble}
\usage{
soft_averaging(predictions_list)
}
\arguments{
\item{predictions_list}{A list of prediction vectors from different models.}
}
\value{
A vector of averaged predictions.
}
\description{
Combines predictions from multiple models by calculating
the average of  predicted probabilties.
}
\examples{
predictions1 <- runif(10, 0, 1)
predictions2 <- runif(10, 0, 1)
avg_predictions <- soft_averaging(list(predictions1, predictions2))
}
